smtapp_core/processors/image_processor.py
"""
Image file processor
"""
from typing import Dict, Any
import os
from PIL import Image

from processors.base_processor import BaseProcessor


class ImageProcessor(BaseProcessor):
    """Processor for image files"""
    
    def supports(self, file_extension: str) -> bool:
        """Check if image file is supported"""
        return file_extension.lower() in ["jpg", "jpeg", "png", "gif", "bmp"]
    
    def process(self, file_path: str) -> Dict[str, Any]:
        """
        Process image file
        Note: OCR would require Tesseract or similar
        
        Args:
            file_path: Path to image file
            
        Returns:
            Dictionary with metadata
        """
        try:
            # Open image
            img = Image.open(file_path)
            
            # Get metadata
            file_ext = file_path.split(".")[-1].lower()
            metadata = {
                "file_type": file_ext,
                "file_size": os.path.getsize(file_path),
                "width": img.width,
                "height": img.height,
                "format": img.format,
                "mode": img.mode,
                "status": "Image OCR not yet implemented"
            }
            
            # Placeholder text
            text = f"[Image file: {os.path.basename(file_path)}]\n"
            text += f"Dimensions: {img.width}x{img.height}\n"
            text += f"Format: {img.format}\n"
            text += "Note: OCR requires additional setup (Tesseract)"
            
            return {
                "text": text,
                "metadata": metadata
            }
            
        except Exception as e:
            raise Exception(f"Error processing image: {str(e)}")

