smtapp_core/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pgvector==0.2.4

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
toml==0.10.2

# File Processing
pypdf==3.17.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
pillow==10.1.0
python-magic==0.4.27

# Audio/Video Processing
pydub==0.25.1
moviepy==1.0.3
opencv-python==4.8.1.78

# ML and NLP
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
langchain==0.0.340
langchain-community==0.0.1

# Ollama client
ollama==0.1.6

# HuggingFace
huggingface-hub==0.19.4

# Vector database
chromadb==0.4.18

# Utilities
httpx==0.25.2
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

